# Peeling last loop iteration (K), changing kPerBlock for the last iteration