--- a/mlir/include/mlir/Dialect/Rock/utility/loweringUtils.h
+++ b/mlir/include/mlir/Dialect/Rock/utility/loweringUtils.h
@@ -65,12 +65,12 @@ RegsAsMatrixSubTiles transposeSubTileViews(PatternRewriter &rewriter,
 // This function will create views of the register buffer of the loaded tile
 // of a matrix in global memory. Those views will provide sub-tiles of the
 // respective hierarchy within the GPU. See above about RegsAsMatrixSubTiles
-FailureOr<RegsAsMatrixSubTiles>
-getLoadRegsAsTileViews(OpBuilder &b, Location loc, Value globalBuffer,
-                       StringRef dName, ArrayRef<StringRef> bidGridOrder,
-                       ArrayRef<int64_t> bidGridLengths, int64_t blockSize,
-                       int64_t kPerBlock, int64_t dPerBlock, int64_t kPerThread,
-                       int64_t dPerThread, bool isKContigousDim);
+FailureOr<RegsAsMatrixSubTiles> getLoadRegsAsTileViews(
+    OpBuilder &b, Location loc, Value globalBuffer, StringRef dName,
+    ArrayRef<StringRef> bidGridOrder, ArrayRef<int64_t> bidGridLengths,
+    int64_t blockSize, int64_t kPerBlock, int64_t dPerBlock, int64_t kPerThread,
+    int64_t dPerThread, bool isKContigousDim, int64_t padK = 0,
+    int64_t prevKperBlock = 0, int64_t effectiveBlockSize = 0);
 
 // This function will create views of the register buffer of the loaded tile
 // but packed as kOuterPerThread, dPerThread and kPackPerThread for max
@@ -81,7 +81,8 @@ FailureOr<RegsAsMatrixSubTiles> getPackedRegsAsTileViews(
     ArrayRef<StringRef> bidGridOrder, ArrayRef<int64_t> bidGridLengths,
     int64_t blockSize, int64_t kPerBlock, int64_t dPerBlock, int64_t kPerThread,
     int64_t dPerThread, int64_t kpack, bool isKContigousDim,
-    bool doSwapThreadIterSubDimsForD = false);
+    bool doSwapThreadIterSubDimsForD = false, int64_t padK = 0,
+    int64_t prevKperBlock = 0, int64_t effectiveBlockSize = 0);
 
 bool isWrWAtomicKernel(GemmFeatures features, Type dataType,
                        bool requiredPadding);