… mlir/lib/Dialect/Rock/Transforms/GridwiseGemmToBlockwise.cpp
… mlir/lib/Dialect/Rock/utility/transformMapUtils.cpp
… mlir/test/Dialect/Rock/gridwise_gemm_accel_lowering.mlir