@pfultz2 2024-10-15
How does this improve the performance for the resnet50 configs?