@manupak 2024-10-16
Im not actively reviewing the PR now as we discussed.

My understanding is if thing are performant, we ll do an initial lowering of : 
```mlir
rock.gridwise_gemm C = A x B
```

to 

```mlir
rock.gridwise_partial_gemm C += A[ : , 0:k1] x B[0:k1 , :]
rock.gridwise_partial_gemm C += A[ : , k1:K] x B[k1:K , :]
rock.threadwise_write_all C 
```

where the code here will be put under lowering of `rock.gridwise_partial_gemm`